refactor: rename icons folder to Icon
