+++ src/components/EditingPanel.tsx
@@ -5,9 +5,9 @@ import TabFilters from "@/components/TabFilters"
 import TabRatio from "@/components/TabRatio"
 import TabTemplate from "@/components/TabTemplate"
 
-import RatioIcon from "@/components/icons/RatioIcon"
-import TemplateIcon from "@/components/icons/TemplateIcon"
-import FiltersIcon from "@/components/icons/FiltersIcon"
+import RatioIcon from "@/components/Icon/RatioIcon"
+import TemplateIcon from "@/components/Icon/TemplateIcon"
+import FiltersIcon from "@/components/Icon/FiltersIcon"
 
 import { SelectedTabType } from "@/types"
 import { useTabData } from "@/lib/hooks/useReduxData"
