+++ src/components/Icon/RatioIcon.tsx
