+++ src/components/Icon/TemplateIcon.tsx
